Rename model_adapter to form_to_dbmodel #16
